save changes to merge

+++ client/src/components/Gallery/CreateGallery/CreateGallery.jsx
@@ -1,8 +1,10 @@
 import React, { useEffect } from 'react';
 import { useState } from 'react';
-import { AiOutlineCloseCircle } from "react-icons/ai";
+import { BiTrash } from "react-icons/bi";
 import './CreateGallery.css';
 import axios from 'axios';
+import Swal from "sweetalert2";
+import validate from './Validations';
 import { useDispatch, useSelector } from 'react-redux';
 import { ClearGallery, deleteImage, getAllImages, getByName, postImage } from '../../../redux/actions';
 
@@ -16,6 +18,7 @@ export default function CreateGallery() {
   const ImageLoading = useSelector((state) => state.rootReducer.ImageLoading);
   
   const [ file, setFile ] = useState(null);
+  const [ error, setError ] = useState(null);
   const [ title, setTitle ] = useState({
     title: ''
   });
@@ -47,11 +50,28 @@ export default function CreateGallery() {
   }
 
   const HandleDelte = async (image) => {
-    dispatch(ClearGallery())
-    const imagedeleted = axios.delete(`http://localhost:3001/gallery/delete?image_id=${image.id}&public_id=${image.public_id}`)
-    Promise.all([imagedeleted]).then(() => {
-      dispatch(getAllImages())
-    });
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#A7D129',
+      cancelButtonColor: 'rgb(43, 43, 44);',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(ClearGallery())
+        const imagedeleted = axios.delete(`http://localhost:3001/gallery/delete?image_id=${image.id}&public_id=${image.public_id}`)
+        Promise.all([imagedeleted]).then(() => {
+          dispatch(getAllImages())
+        });
+        Swal.fire(
+          'Deleted!',
+          'Your image has been deleted.',
+          'success'
+        )
+      }
+    })
   }
 
   const handleChange = (e) => {
@@ -63,20 +83,30 @@ export default function CreateGallery() {
   }
 
   const HandlerCLick = async () => {
-    if(!file) {
-      return alert('you must chose a file')
+    if(!file || !title.title) {
+      Swal.fire({
+        title: 'There is an error',
+        text: "Title and file are both required",
+        icon: 'warning',
+        showCancelButton: false,
+        showConfirmButton: true,
+        confirmButtonColor: '#A7D129',
+        cancelButtonColor: '#A7D129',
+        confirmButtonText: ' Okey '
+      })
+    }else{
+      const formData = new FormData();
+      formData.append('image', file);
+      dispatch(ClearGallery())
+      setGallery(true)
+  
+      await axios.post(`http://localhost:3001/gallery/post?title=${title.title}`,formData)
+      dispatch(getAllImages())
+      
+      // document.getElementById('floatingInput2').value = null;
+      setFile(null)
     }
     
-    const formData = new FormData();
-    formData.append('image', file);
-    dispatch(ClearGallery())
-    setGallery(true)
-
-    await axios.post(`http://localhost:3001/gallery/post?title=${title.title}`,formData)
-    dispatch(getAllImages())
-    
-    // document.getElementById('floatingInput2').value = null;
-    setFile(null)
   }
 
 
@@ -130,7 +160,7 @@ export default function CreateGallery() {
                           {
                             auth.loggedIn && auth.currentUser.is_admin ? 
                             <div onClick={()=>HandleDelte(img)} className='btn_delete_image' >
-                              <AiOutlineCloseCircle className='tarea-icono' />
+                              <BiTrash className='tarea-icono' />
                             </div>
                              : null 
                           }
@@ -146,7 +176,7 @@ export default function CreateGallery() {
                           {
                             auth.loggedIn && auth.currentUser.is_admin ? 
                             <div onClick={()=>HandleDelte(img)} className='btn_delete_image' >
-                              <AiOutlineCloseCircle className='tarea-icono' />
+                              <BiTrash className='tarea-icono' />
                             </div>
                              : null
                           }
@@ -162,7 +192,7 @@ export default function CreateGallery() {
                           {
                             auth.loggedIn && auth.currentUser.is_admin ? 
                             <div onClick={()=>HandleDelte(img)} className='btn_delete_image' >
-                              <AiOutlineCloseCircle className='tarea-icono' />
+                              <BiTrash className='tarea-icono' />
                             </div>
                              : null
                           }
